requirements.txt
asgiref==3.5.2
cachetools==5.2.0
certifi==2022.6.15
charset-normalizer==2.1.0
Django==4.0.6
django-filter==22.1
djangorestframework==3.13.1
google-api-core==2.8.2
google-api-python-client==2.52.0
google-apis-oauth-django==1.0.0
google-auth==2.9.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.5.2
googleapis-common-protos==1.56.3
httplib2==0.20.4
idna==3.3
oauth2client==4.1.3
oauthlib==3.2.0
protobuf==4.21.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyparsing==3.0.9
pytz==2022.1
requests==2.28.1
requests-oauthlib==1.3.1
rsa==4.8
six==1.16.0
sqlparse==0.4.2
tzdata==2022.1
uritemplate==4.1.1
urllib3==1.26.9
